requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
youtube-search-python
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
faiss-cpu
duckduckgo-search